image_service/requirements.txt
pyzmq
Pillow